Revert "Crud setup"
