+++ flask_app/app/app.py
@@ -3,8 +3,7 @@ The entry point into the flask app
 """
 
 
-from flask import Flask, session, request, redirect, url_for, abort, \
-    render_template, flash
+from flask import Flask
 
 app = Flask(__name__)
 app.config.from_object(__name__)
@@ -23,7 +22,6 @@ def index():
     """
     The home/index page shows signup and signin forms
     """
-    return render_template('index.html')
 
 
 @app.route('/signup', methods=['POST'])
@@ -33,52 +31,13 @@ def signup():
     the signup form on the home/index page
     """
 
-@app.route('/signout')
-def signout():
-    """
-    The signout route logs out the user
-    """
-    error = None
-    if session['user']:
-        user = session['user']
-        user['logged_in'] = False
-        session['user'] = user
-        flash('You have logged out successfully')
-        return redirect(url_for('index'))
-    else:
-        error = 'User not found'
-    return render_template('index.html', error=error)
-
 
 @app.route('/signin', methods=['POST'])
 def signin():
     """
     The signin route handles POST data sent 
     from the signin form on the home/index page
-
-    user registered in session is a 
-    dictionary of username, password and logged_in
     """
-    error = None
-    try:
-        saved_user = session['user']
-    except KeyError:
-        error = 'an error occured in your session. Try refreshing'
-        return render_template('index.html', error=error)
-    
-    if saved_user['logged_in']:
-        error = 'Already logged in'
-    elif request.form['username'] != saved_user['username']:
-        error = 'Invalid username'
-    elif request.form['password'] != saved_user['password']:
-        error = 'Invalid password'
-    else:
-        saved_user['logged_in'] = True
-        session['user'] = saved_user
-        flash('Login successful')
-        return redirect(url_for('show_user_record',
-                     username=saved_user['username']))
-    return render_template('index.html', error=error)
 
 
 @app.route('/user')
@@ -89,13 +48,13 @@ def show_users():
     """
 
 
-@app.route('/user/<str:username>',
+@app.route('/user/<int:user_id>',
  methods=['POST', 'GET'])
-def show_user_record(username):
+def show_user_record(user_id):
     """
     The show_user_record route shows(GET) 
-    the basic details about the user of the said
-    username and the list of shopping lists
+    the basic details about the user of id,
+    user_id and the list of shopping lists
      belonging to that user. 
     
     It allows creation (POST) of new shopping
@@ -103,12 +62,12 @@ def show_user_record(username):
     """
 
 
-@app.route('/user/<str:username>/shoppinglist/<str:title>',
+@app.route('/user/<int:user_id>/shoppinglist/<int:list_id>',
 methods=['POST', 'GET', 'DELETE', 'PUT'])
-def show_single_shoppinglist(username, title):
+def show_single_shoppinglist(user_id, list_id):
     """
-    This view shows(GET) all the items in a shopping list of the 
-    given title belonging to user of the mentioned username. 
+    This view shows(GET) all the items in a shopping list of id 
+    list_id belonging to user user_id. 
 
     It also allows for editting(PUT) and deleting(DELETE) of a 
     shopping list.
@@ -117,9 +76,9 @@ def show_single_shoppinglist(username, title):
     """
 
 
-@app.route('/user/<str:username>/shoppinglist/<str:title>\
-/item/<str:name>', methods=['DELETE', 'PUT'])
-def edit_shopping_item(username, title, name):
+@app.route('/user/<int:user_id>/shoppinglist/<int:list_id>\
+/item/<int:item_id>', methods=['DELETE', 'PUT'])
+def edit_shopping_item(user_id, list_id, item_id):
     """
     This route deals allows deleting (DELETE) or editing(PUT) of an
     item on a list
