+++ flask_app/app/classes/utilities.py
@@ -5,25 +5,15 @@ This module has functions used across different classes
 import re
 
 
-def check_type(obj, type_object, *args, error_string='Invalid type'):
+def check_type(obj, type_object, error_string='Invalid type'):
     """
     Checks the type of obj against the type_object
     and returns True is the same or else raises TypeError
     """
-    if not isinstance(type_object, type):
-        raise ValueError('second argument of check_type should be\
-        a type not a %s')
-
     if isinstance(obj, type_object):
             return True
-
-    arg_length = len(args)
-    if arg_length > 0:
-        for item in args:
-            if isinstance(obj, item):
-                return True
-
-    raise TypeError(error_string)   
+    else:
+        raise TypeError(error_string)
 
 
 def check_email_format(email):
@@ -36,12 +26,12 @@ def check_email_format(email):
     return True
 
 
-def check_password_format(password, min_length=6):
+def check_password_format(password):
     """
     Checks to ensure that the password is a string of not less
-    than min_length
+    than 6 characters
     """
     if check_type(password, str):
-        if len(password) < min_length:
+        if len(password) < 6:
             raise ValueError("Your password is too short")
         return True
