--- flask_app/app/crud/shopping_item.py
@@ -1,31 +1,0 @@
-"""
-This file contains functionality for
-- create, read, update and delete shopping items
-"""
-
-
-def create_shopping_item(shopping_list, **kwargs):
-    """
-    Create an item in shopping list for the user
-    """
-
-def update_shopping_item(shopping_list, item, **kwargs):
-    """
-    Update an item in a shopping list for the user
-    """
-
-
-def delete_shopping_item(shopping_list, item):
-    """
-    delete an item in a shopping list for the user
-    """
-
-def list_shopping_items(shopping_list):
-    """
-    returns all items in a shopping list for a given user
-    """
-
-def single_shopping_item(shopping_list, item):
-    """
-    returns a single item of a shopping_list about a shopping_list
-    """
