--- flask_app/app/crud/shopping_list.py
@@ -1,34 +1,0 @@
-"""
-This file contains the following functionality
-- create, read, update and delete shopping list
-"""
-
-
-def create_shopping_list(user, **kwargs):
-    """
-    Create a shopping list for the user
-    """
-
-
-def update_shopping_list(user, **kwargs):
-    """
-    Update a shopping list for the user
-    """
-
-
-def delete_shopping_list(user, shopping_list):
-    """
-    delete a shopping list for the user
-    """
-
-
-def list_shopping_lists(user):
-    """
-    returns all shopping lists for a given user
-    """
-
-
-def single_shopping_list(user, shopping_list):
-    """
-    returns the appropriate data about a shopping_list
-    """
